+++ src/stats/hexbin.jl
@@ -54,8 +54,9 @@ function data_limits(hb::Hexbin)
     nw = widths(bb) .+ (ms..., 0.0f0)
     no = bb.origin .- ((ms ./ 2.0f0)..., 0.0f0)
 
-    return Rect3f(no, nw)
+    return Rect3d(no, nw)
 end
+boundingbox(p::Hexbin) = transform_bbox(p, data_limits(hb))
 
 get_weight(weights, i) = Float64(weights[i])
 get_weight(::StatsBase.UnitWeights, i) = 1e0